Moved theme provider
