+++ src/presentation/apps/sidebar/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import { useContext } from "react";
+import { useTheme } from "../../providers/theme_provider.tsx";
 import { ActionButton } from "./components/actionButton.tsx";
 import { ReloadButton } from "./components/reloadButton.tsx";
 import { ModalPopup } from "./components/modalPopup.tsx";
@@ -8,41 +8,45 @@ import { DeclutterHeader } from "./components/header.tsx";
 import { ModalProvider } from "./providers/modalContext.tsx";
 import ThemeToggle from "./components/themeToggle.tsx";
 import { AppProvider } from "../../providers/app_provider.tsx";
-import { ThemeProvider } from "../../providers/ThemeProvider.tsx";
+import { ThemeProvider } from "../../providers/theme_provider.tsx";
 import { ActionsProvider } from "../../../_shared/providers/actionsContext.tsx";
-import { ThemeContext } from "../../providers/themeContext.ts";
 
 function App() {
-  const { theme } = useContext(ThemeContext);
-
   return (
     <AppProvider>
       <ActionsProvider>
         <ThemeProvider>
-          <ModalProvider>
-            <div id="declutter-body" className={theme}>
-              <DeclutterHeader />
-
-              <div className="button-bar">
-                <div className="sender-actions">
-                  <ActionButton id="unsubscribe-button" />
-                  <ActionButton id="delete-button" />
-                </div>
-
-                <div style={{ display: "flex" }}>
-                  <ReloadButton />
-                  <ThemeToggle />
-                </div>
-              </div>
-
-              <SendersContainer />
-
-              <ModalPopup />
-            </div>
-          </ModalProvider>
+          <AppWithTheme />
         </ThemeProvider>
       </ActionsProvider>
     </AppProvider>
+  );
+}
+
+function AppWithTheme() {
+  const { theme } = useTheme();
+  return (
+    <ModalProvider>
+      <div id="declutter-body" className={theme}>
+        <DeclutterHeader />
+
+        <div className="button-bar">
+          <div className="sender-actions">
+            <ActionButton id="unsubscribe-button" />
+            <ActionButton id="delete-button" />
+          </div>
+
+          <div style={{ display: "flex" }}>
+            <ReloadButton />
+            <ThemeToggle />
+          </div>
+        </div>
+
+        <SendersContainer />
+
+        <ModalPopup />
+      </div>
+    </ModalProvider>
   );
 }
