+++ src/presentation/apps/sidebar/main.tsx
@@ -5,6 +5,6 @@ import App from "./App.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-      <App />
+    <App />
   </StrictMode>,
 );
