+++ src/presentation/apps/tutorial/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "../../../index.css";
 import Tutorial from "./Tutorial.tsx";
-import { ThemeProvider } from "../../providers/ThemeProvider.tsx";
+import { ThemeProvider } from "../../providers/theme_provider.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
